Fix usb.c/h build errors: type mismatch, missing OS macros and memset
- usb_poll return type: unsigned long -> u32 to match implementation
- Add PR/os.h for OS_USEC_TO_CYCLES and ALIGN macros
- Declare memset since decomp libc string.h omits it

+++ lib/netplay/usb.h
@@ -79,7 +79,7 @@
         @return The data header, or 0
     ==============================*/
     
-    extern unsigned long usb_poll(void);
+    extern u32 usb_poll(void);
     
     
     /*==============================
